requirements.txt
Pillow
fpdf
matplotlib
requests
qrcode
sympy
pandas
